bot-google-ads/bot/requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.27.0
mistralai==0.0.12
pyodbc==5.0.1
beautifulsoup4==4.12.3
requests==2.31.0
pydantic==2.5.3
python-dotenv==1.0.0
lxml==5.1.0
google-ads==24.1.0
